+++ cli/src/cols.rs
@@ -24,6 +24,9 @@ impl Cols {
     pub fn is_empty(&self) -> bool {
         self.0.is_empty()
     }
+    pub fn len(&self) -> usize {
+        self.0.len()
+    }
     pub fn contains(
         &self,
         tbl: Col,